COMMERCE-6038 talend-common - Reorganize - HeadlessUtil - specialized for liferay headless api domain
